Diagram: Removed dependency to the tooltips plugin
There are now different types of tooltips for the ports. The diagram
language itself contains only text tooltips and supports them.
There is a new language, extending the diagram language, that allows
editor cells as tooltips. It has a dependency on the tooltips language.

There are now two plugin. A "diagram" and a "diagram.base" plugin. The
diagram.base plugin is supposed to have no dependencies on other plugins.
The diagram plugin provides extensions to the base plugin.